revert file changes to tests
